requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
tensorflow-cpu
joblib
python-multipart